requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.98
google-generativeai==0.8.4
PyAudio==0.2.14
